% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bstickPC.R
\name{bstickPC}
\alias{bstickPC}
\title{Broken stick analysis}
\usage{
bstickPC(pcaOutput)
}
\arguments{
\item{pcaOutput}{object assigned to output of pcaPlot()}
}
\value{
scree plot from pcaPlot() with dotted red line indicating suggested importance cutoff
}
\description{
This code runs Broken Stick method for Principal Component selection
}
