% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbowPC.R
\name{elbowPC}
\alias{elbowPC}
\title{Elbow Method}
\usage{
elbowPC(pcaOutput)
}
\arguments{
\item{pcaOutput}{object assigned to the output of pcaPlot()}
}
\value{
scree plot from pcaPlot() with suggested cutoff marked by a dotted red line
}
\description{
This function performs Elbow Method for Principal Component importance
}
