% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodseqSetup.R
\name{foodseqSetup}
\alias{foodseqSetup}
\title{FoodSeq Setup Pipeline}
\usage{
foodseqSetup(
  physeq,
  amplicon = "",
  sepVar = NULL,
  collapse = FALSE,
  CommonNames = NULL,
  CommonNameCol = NULL
)
}
\arguments{
\item{physeq}{raw phyloseq object}

\item{amplicon}{trnl or 12s}

\item{sepVar}{define sam_data variable by which CLR transform will be done separately (e.g., seq_date, Study)}

\item{collapse}{optional to run collapseNoMismatch()}

\item{CommonNames}{optional space to add common names file}

\item{CommonNameCol}{optional to add name of common name column to make a new column that's labeled with common name if available and lowestLevel if not}
}
\value{
ps.raw = phyloseq object with no changes from physeq other than adding p/aMR, Shannon, and reads to samdf (no NA/human reads removed)

ps = phyloseq object with no changes from physeq other than changes to ps.raw + animals glommed + NA/human reads removed

ps.ra = changes to ps (but NA's conserved) + transformed to relative abundance

ps.filt.clr = changes to ps + clr transformation + NA/human reads removed

ps.na = phyloseq object sub-setted for samples containing NA reads
}
\description{
This function runs raw phyloseq object through typical inital processing steps.
}
