% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naCheck.R
\name{naCheck}
\alias{naCheck}
\title{NA Checker}
\usage{
naCheck(physeq)
}
\arguments{
\item{physeq}{phyloseq object of interest}
}
\value{
for each study in the phyloseq object, returns the number of NA's present
}
\description{
Checks a phyloseq object for any taxa that have no superkingdom assignment
}
