% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paranPC.R
\name{paranPC}
\alias{paranPC}
\title{Parallel Analysis}
\usage{
paranPC(physeq, pcaOutput, B = 1000, centile = 95, seed = 123)
}
\arguments{
\item{pcaOutput}{object storing the output from pcaPlot()}

\item{B}{number of permutations}

\item{centile}{percentile threshold}

\item{seed}{random seed for reproducibility}

\item{physe}{filtered and clr transformed phyloseq object on which pcaPlot() was run}
}
\value{
cutoff_pc = last significantly important pc

var_explained = variance explained from PC1-cutoff_pc

paran.df = data frame used to generate plot

paran.plot = parallel analysis plot (scree table with cutoff)

qc.df = data frame with prcomp() numbers from internal and external (pcaPlot()) pca generation

qc.plot = scatterplot of eigenvalues from internal and external (pcaPlot()) pca generation
}
\description{
Runs parallel analysis (permutation) for principal component importance
}
