% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaProject.R
\name{pcaProject}
\alias{pcaProject}
\title{PCA Projection}
\usage{
pcaProject(new_ps, old_ps, pca_output)
}
\arguments{
\item{new_ps}{new phyloseq object; with NA's, non-clr transformed}

\item{old_ps}{old phyloseq object; with NA's, non-clr transformed}

\item{pca_output}{pcaPlot()$pca_output}
}
\value{
pseudocount = number used when running clr transformation

ps.filt.clr = new_ps after going through clr trasnformation exactly the same as old_ps

projection.scores = data frame with PC scores for each sample in new_ps
}
\description{
projects new data into existing PCA space
}
