% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcCorr.R
\name{pcCorr}
\alias{pcCorr}
\title{PC correlations}
\usage{
pcCorr(data, xVar, yVar, colorVar, continuous = TRUE, mid = 50)
}
\arguments{
\item{data}{pcaPlot()$pca.df}

\item{xVar}{variable name of interest (e..g, PC1)}

\item{yVar}{variable name of interest (e.., age)}

\item{colorVar}{variable by which to color points}

\item{continuous}{T/F}

\item{mid}{determines midpoint of gradient color scale}
}
\description{
makes scatterplot with PC and variable of interest
}
