% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcDens.R
\name{pcDens}
\alias{pcDens}
\title{Density Plot}
\usage{
pcDens(
  data,
  x = "PC1",
  y = "country",
  title = NULL,
  scale = 2,
  alpha = 0.7,
  title.size = 14,
  text.size = 12,
  customColors = NULL
)
}
\arguments{
\item{data}{pcaPlot() output}

\item{x}{"PC" of interest}

\item{y}{variable of interest (e.g., country)}

\item{title}{optional input of title}

\item{scale}{= 2}

\item{alpha}{= 0.7}

\item{title.size}{= 14}

\item{text.size}{= 12}

\item{customColors}{optional color palette inclusion}
}
\value{
density plot
}
\description{
generates density plot for two variables of interest
}
