% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psPerm.R
\name{psPerm}
\alias{psPerm}
\title{Phyloseq PERMANOVA}
\usage{
psPerm(
  physeq,
  equation,
  method = "euclidean",
  by = "terms",
  permutations = 1000,
  strata = NULL,
  seed = 123
)
}
\arguments{
\item{physeq}{filtered and clr transfomed phyloseq object}

\item{equation}{e.g., "country + age + treatment" or "country*age + treatment"}

\item{method}{euclidean}

\item{by}{"terms" for sequential assessment/control for covariates in order, "margins" for individual assessment/control for all other variables}

\item{permutations}{number of permutations}

\item{strata}{grouping variable}

\item{seed}{random seed for reproducibility}
}
\value{
PERMANOVA output
}
\description{
function to run permanova on foodseq phyloseq object
}
