% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remNAcol.R
\name{remNAcol}
\alias{remNAcol}
\title{Remove NA Columns}
\usage{
remNAcol(ps)
}
\arguments{
\item{ps}{phyloseq object with sample data table}
}
\value{
ps phyloseq object with NA columns removed
}
\description{
removes columns with all NA rows from the sample data table of a phyloseq object
}
