% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxPrev.R
\name{taxPrev}
\alias{taxPrev}
\title{Specific Taxa Prevalence}
\usage{
taxPrev(
  physeq,
  taxList,
  name = "ShortName",
  facet = NULL,
  nrow = NULL,
  title = NULL,
  taxListTitle = NULL,
  color = TRUE,
  localCol = TRUE,
  ylim100 = TRUE,
  labWidth = 60,
  numLab = FALSE,
  titleSize = 16,
  axtitleSize = 16,
  textSize = 12,
  stripSize = 10
)
}
\arguments{
\item{physeq}{phyloseq object}

\item{taxList}{list of asv's of interest}

\item{name}{name of common name column in tax table of phyloseq object}

\item{facet}{name of samdf column by which to facet (optional)}

\item{nrow}{number of rows for facet_wrap()}

\item{title}{optional title of plot}

\item{taxListTitle}{optional title for taxList (y axis title)}

\item{color}{option for graph to be colored}

\item{localCol}{option for colors to be assigned locally versus globally (e.g., color by rank vs taxa)}

\item{ylim100}{option for setting ylim(0,100)}

\item{labWidth}{option to set character breakpoint for label wrapping}

\item{numLab}{option to add value labels to bar chart}

\item{titleSize}{set plot title size}

\item{axtitleSize}{set axis title size}

\item{textSize}{set plot text size}

\item{stripSize}{set strip text size}

\item{stripText}{optional labels for strip text}
}
\value{
df  = data frame of taxa and prevalence

plot  = bar chart of taxa and prevalence
}
\description{
Gives the prevalence of a particular list of taxa (asv's) in a phyloseq object. Optional faceting.
}
