% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topTaxaPrevalence.R
\name{topTaxaPrevalence}
\alias{topTaxaPrevalence}
\title{Top Taxa Prevalence}
\usage{
topTaxaPrevalence(phyloseq_obj, n, title, amplicon)
}
\arguments{
\item{phyloseq_obj}{phyloseq of interest}

\item{n}{number of taxa to plot}

\item{title}{optional title of plot}

\item{amplicon}{trnl or 12s}
}
\value{
bar plot of top n taxa

top.taxa = data frame of top taxa
}
\description{
finds and plots the top n most prevalent taxa in a phyloseq object
}
