% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapLabels.R
\name{wrapLabels}
\alias{wrapLabels}
\title{Wrap Labels}
\usage{
wrapLabels(x, width = 60)
}
\arguments{
\item{x}{string of interest}

\item{width}{character length for breakpoint}
}
\value{
input string with new line at every width characters
}
\description{
Takes in a string and adds line breaks at specified breakpoints
}
