% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodseqSetup.R
\name{foodseqSetup}
\alias{foodseqSetup}
\title{FoodSeq Setup Pipeline}
\usage{
foodseqSetup(physeq, amplicon = "", collapse = FALSE)
}
\arguments{
\item{physeq}{raw phyloseq object}

\item{amplicon}{trnl or 12s}

\item{collapse}{optional to run collapseNoMismatch()}
}
\value{
ps.raw = phyloseq object with no changes from physeq other than adding p/aMR, Shannon, and reads to samdf (no NA/human reads removed)

ps = phyloseq object wiht no changes from physeq other than changes to ps.raw + animals glommed

ps.ra = changes to ps + transformed to relative abundance

ps.filt.clr = changes to ps + NA/human reads removed + clr transformation
}
\description{
This function runs raw phyloseq object through typical inital processing steps.
}
