% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaPlot.R
\name{pcaPlot}
\alias{pcaPlot}
\title{\subsection{PCA plot and biplot
Code written by Ben, adapted by Caroline}{
}}
\usage{
pcaPlot(
  ps,
  colorVar = NULL,
  colorName = NULL,
  nTaxa = 10,
  customColors = NULL,
  customGradient = NULL,
  mid = "mean",
  xPC = 1,
  yPC = 2,
  ellipse = FALSE
)
}
\arguments{
\item{ps}{clr transformed and filtered data}

\item{colorVar}{variable from samdf to color samples by}

\item{colorName}{what to display variable name as in legend}

\item{nTaxa}{number of taxa to display}

\item{customColors}{optional named vector of colors}

\item{customGradient}{optional df of low/high colors for gradient}

\item{mid}{for customGradient -- midpoint to median, mean, middle of range}

\item{xPC}{Principal Component for x-axis}

\item{yPC}{Principal Component for y-axis}

\item{ellipse}{optional add centroid ellipses}
}
\value{
pca biplot with nTaxa factor loadings displayed

samdf with all pc's added as columns

data frame with loadings for pca space

scree plot

table with scree plot numbers

output of prcomp()
}
\description{
This function runs a Principal Component Analysis.
}
