% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_n_taxa.R
\name{top_n_taxa}
\alias{top_n_taxa}
\title{Top Taxa}
\usage{
top_n_taxa(
  physeq,
  n = 10,
  name = ShortName,
  title = NA,
  remNA = FALSE,
  color = TRUE
)
}
\arguments{
\item{physeq}{phyloseq object}

\item{n}{number of taxa to plot}

\item{name}{name of common name column (e.g., ShortName)}

\item{title}{optional title}

\item{remNA}{option to keep/remove taxa that don't have common name assignment}

\item{color}{option to add/remove color to the graph}
}
\value{
top_taxa = data frame of top n taxonomy table

top_taxa_plot = bar graph of top n taxa
}
\description{
finds and plots top taxa from a phyloseq
}
